dockerstreamlit/requirements.txt
pandas==2.2.1
streamlit==1.30.0
mysql-connector-python==8.3.0